Refactored BMP280 driver
The driver now implements the PressureSensor interface
